Add Note mentioning to drain caps.
Adding in "5. Try to disconnect the power cable and switch the button on the power supply to off. Following that, please try to hold down the systems power button 3 times for 10 seconds each. This will drain capacitors and can potentially allow the system to power up normally."

Just had a call from someone with a Thelio that was not powering on. He said he followed all instructions in the article. I reviewed the article and found this step was not included. I'd like to add it. I think it can help.